ScanNet/requirements.txt
absl-py==1.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
biopython==1.78
bleach==4.1.0
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastapi==0.83.0
Flask==2.0.3
gast==0.6.0
google-pasta==0.2.0
grpcio==1.48.2
h11==0.13.0
h5py==2.10.0
idna==3.10
immutables==0.19
importlib-metadata==4.8.3
ipydatawidgets==4.3.1.post1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywebrtc==0.5.0
ipywidgets==7.5.1
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.6.0
notebook==6.4.10
numba==0.52.0
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.1
parso==0.7.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.9.2
Pygments==2.14.0
pyparsing==3.1.4
pyrsistent==0.18.0
python-dateutil==2.9.0.post0
pythreejs==2.3.0
pytz==2025.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.3
six==1.17.0
sniffio==1.2.0
starlette==0.19.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==4.3.3
traittypes==0.2.1
typing_extensions==4.1.1
urllib3==1.26.20
uvicorn==0.16.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wrapt==1.16.0
zipp==3.6.0
